Move to BackgroundJob
* No need to have a separate class that only calls the BackgroundScanner
* Added repairstep to remove old task
* Bumped version
